functions to create a prompt and confirm prompts for each criteria.
Using the provided script.js code, I am creating a generatePassword function to start the prompts - within this function it will call the getPasswordLength function to return the value of the password length. Then I have created confirm prompts for the Special characters to return a true or false value and so on for each of the passwords criteria.