For Loop to test password Values and output the password
Creating a for loop that takes in the values of the true/false from the passwordValues prompts - that tests if it is upper, lower, number, or special. If it is any of those values then it will input one element from the array at random. This will return a randomly generated password with the selected criteria. Added in screenshots and links :)